+++ applications/fishing-map/src/features/app/App.tsx
@@ -8,7 +8,9 @@ import { MapContext } from 'features/map/map-context.hooks'
 import useDebugMenu from 'features/debug/debug.hooks'
 import {
   isWorkspaceLocation,
+  selectLocationCategory,
   selectLocationType,
+  selectUrlTimeRange,
   selectUrlViewport,
   selectWorkspaceId,
 } from 'routes/routes.selectors'
@@ -28,17 +30,19 @@ import { AsyncReducerStatus } from 'utils/async-slice'
 import useViewport, { useMapFitBounds } from 'features/map/map-viewport.hooks'
 import { selectIsAnalyzing } from 'features/analysis/analysis.selectors'
 import { isUserLogged } from 'features/user/user.selectors'
-import { DEFAULT_WORKSPACE_ID } from 'data/workspaces'
+import { DEFAULT_WORKSPACE_ID, WorkspaceCategories } from 'data/workspaces'
 import { HOME, WORKSPACE, USER, WORKSPACES_LIST } from 'routes/routes'
 import { fetchWorkspaceThunk } from 'features/workspace/workspace.slice'
 import { t } from 'features/i18n/i18n'
+import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
+import Welcome from 'features/welcome/Welcome'
 import { useAppDispatch } from './app.hooks'
 import { selectAnalysisQuery, selectSidebarOpen } from './app.selectors'
 import styles from './App.module.css'
 import { useAnalytics } from './analytics.hooks'
 
-const Map = lazy(() => import('features/map/Map'))
-const Timebar = lazy(() => import('features/timebar/Timebar'))
+const Map = lazy(() => import(/* webpackChunkName: "Map" */ 'features/map/Map'))
+const Timebar = lazy(() => import(/* webpackChunkName: "Timebar" */ 'features/timebar/Timebar'))
 
 declare global {
   interface Window {
@@ -58,6 +62,9 @@ const Main = () => {
   )
 }
 
+const MARINE_MANAGER_LAST_VISIT = 'MarineManagerLastVisit'
+const isFirstTimeVisit = !localStorage.getItem(MARINE_MANAGER_LAST_VISIT)
+
 function App(): React.ReactElement {
   useAnalytics()
   const dispatch = useAppDispatch()
@@ -68,15 +75,29 @@ function App(): React.ReactElement {
   const workspaceLocation = useSelector(isWorkspaceLocation)
   const isAnalysing = useSelector(selectIsAnalyzing)
   const narrowSidebar = workspaceLocation && !analysisQuery
+  const { debugActive, dispatchToggleDebugMenu } = useDebugMenu()
+
+  const locationIsMarineManager =
+    useSelector(selectLocationCategory) === WorkspaceCategories.MarineManager
+  const [welcomePopupOpen, setWelcomePopupOpen] = useState(
+    locationIsMarineManager && isFirstTimeVisit
+  )
+  useEffect(() => {
+    if (locationIsMarineManager)
+      localStorage.setItem(MARINE_MANAGER_LAST_VISIT, new Date().toISOString())
+  }, [locationIsMarineManager])
 
   const fitMapBounds = useMapFitBounds()
   const { setMapCoordinates } = useViewport()
+  const { setTimerange } = useTimerangeConnect()
 
   const locationType = useSelector(selectLocationType)
   const currentWorkspaceId = useSelector(selectCurrentWorkspaceId)
   const workspaceCustomStatus = useSelector(selectWorkspaceCustomStatus)
+  const showToggle = useSelector(isWorkspaceLocation)
   const userLogged = useSelector(isUserLogged)
   const urlViewport = useSelector(selectUrlViewport)
+  const urlTimeRange = useSelector(selectUrlTimeRange)
   const urlWorkspaceId = useSelector(selectWorkspaceId)
 
   // TODO review this as is needed in analysis and workspace but adds a lot of extra logic here
@@ -94,6 +115,12 @@ function App(): React.ReactElement {
         if (!urlViewport && resolvedAction.payload?.viewport) {
           setMapCoordinates(resolvedAction.payload.viewport)
         }
+        if (!urlTimeRange && resolvedAction.payload?.startAt && resolvedAction.payload?.endAt) {
+          setTimerange({
+            start: resolvedAction.payload?.startAt,
+            end: resolvedAction.payload?.endAt,
+          })
+        }
       }
       actionResolved = true
     }
@@ -102,6 +129,8 @@ function App(): React.ReactElement {
       workspaceCustomStatus !== AsyncReducerStatus.Loading &&
       (homeNeedsFetch || hasWorkspaceIdChanged)
     ) {
+      // TODO Can we arrive in a situation where no workspace is ever loaded?
+      // In that case static timerange will need to be set manually
       fetchWorkspace()
     }
     return () => {
@@ -122,8 +151,6 @@ function App(): React.ReactElement {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
-
-  const { debugActive, dispatchToggleDebugMenu } = useDebugMenu()
 
   useEffect(() => {
     dispatch(fetchUserThunk())
@@ -155,6 +182,7 @@ function App(): React.ReactElement {
       <Suspense fallback={null}>
         <SplitView
           isOpen={sidebarOpen}
+          showToggle={showToggle}
           onToggle={onToggle}
           aside={<Sidebar onMenuClick={onMenuClick} />}
           main={<Main />}
@@ -177,6 +205,17 @@ function App(): React.ReactElement {
       >
         <DebugMenu />
       </Modal>
+      {welcomePopupOpen && (
+        <Suspense fallback={null}>
+          <Modal
+            header={false}
+            isOpen={welcomePopupOpen}
+            onClose={() => setWelcomePopupOpen(false)}
+          >
+            <Welcome />
+          </Modal>
+        </Suspense>
+      )}
     </MapContext.Provider>
   )
 }