+++ applications/fishing-map/src/features/i18n/i18n.ts
@@ -44,6 +44,10 @@ i18n
     },
   })
 
+i18n.on('languageChanged', (lng) => {
+  document.documentElement.setAttribute('lang', lng)
+})
+
 const t = i18n.t.bind(i18n)
 
 export { t }