+++ applications/fishing-map/src/features/timebar/timebar.hooks.ts
@@ -1,30 +1,60 @@
 import { useDispatch, useSelector } from 'react-redux'
-import { useCallback, useEffect } from 'react'
+import { useCallback, useEffect, useMemo } from 'react'
+import { atom, useRecoilState } from 'recoil'
+import { debounce } from 'lodash'
 import { TimebarVisualisations } from 'types'
-import { selectTimeRange, selectTimebarVisualisation } from 'features/app/app.selectors'
+import { selectTimebarVisualisation } from 'features/app/app.selectors'
 import { useLocationConnect } from 'routes/routes.hook'
 import {
   selectActiveActivityDataviews,
   selectActiveTrackDataviews,
 } from 'features/dataviews/dataviews.selectors'
-import { setStaticTime, selectHasChangedSettingsOnce, changeSettings } from './timebar.slice'
+import store, { RootState } from 'store'
+import { updateUrlTimerange } from 'routes/routes.actions'
+import { selectUrlTimeRange } from 'routes/routes.selectors'
+import { selectHasChangedSettingsOnce, changeSettings, Range } from './timebar.slice'
+
+export const TimeRangeAtom = atom<Range | null>({
+  key: 'timerange',
+  default: null,
+  effects_UNSTABLE: [
+    ({ trigger, setSelf, onSet }) => {
+      const urlTimeRange = selectUrlTimeRange(store.getState() as RootState)
+      const dispatch = useDispatch()
+
+      if (trigger === 'get' && urlTimeRange) {
+        setSelf({
+          ...urlTimeRange,
+        })
+      }
+      const updateTimerangeDebounced = debounce(dispatch(updateUrlTimerange), 1000)
+      onSet((timerange) => {
+        if (timerange) {
+          updateTimerangeDebounced({ ...timerange })
+        }
+      })
+    },
+  ],
+})
 
 export const useTimerangeConnect = () => {
-  const { dispatchQueryParams } = useLocationConnect()
-  const dispatch = useDispatch()
-  const { start, end } = useSelector(selectTimeRange)
-  // TODO needs to be debounced like viewport
-  const dispatchTimeranges = useCallback(
-    (event: { start: string; end: string; source: string }) => {
-      const range = { start: event.start, end: event.end }
-      if (event.source !== 'ZOOM_OUT_MOVE') {
-        dispatch(setStaticTime(range))
-      }
-      dispatchQueryParams(range)
+  const [timerange, setTimerange] = useRecoilState(TimeRangeAtom)
+
+  const onTimebarChange = useCallback(
+    (start: string, end: string) => {
+      setTimerange({ start, end })
     },
-    [dispatchQueryParams, dispatch]
+    [setTimerange]
   )
-  return { start, end, dispatchTimeranges }
+  return useMemo(() => {
+    return {
+      start: timerange?.start,
+      end: timerange?.end,
+      timerange,
+      setTimerange,
+      onTimebarChange,
+    }
+  }, [onTimebarChange, timerange, setTimerange])
 }
 
 export const useTimebarVisualisation = () => {