+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.50
+
+### Patch Changes
+
+- Updated dependencies [1a3d0960]
+  - @globalfishingwatch/dataviews-client@7.0.0
+  - @globalfishingwatch/layer-composer@7.0.0
+  - @globalfishingwatch/timebar@1.6.0
+
+## 1.1.49
+
+### Patch Changes
+
+- Updated dependencies [7d060771]
+  - @globalfishingwatch/api-client@1.7.1
+  - @globalfishingwatch/dataviews-client@6.5.3
+  - @globalfishingwatch/layer-composer@6.1.4
+
 ## 1.1.48
 
 ### Patch Changes