+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.85
+
+### Patch Changes
+
+- Updated dependencies [37935df0]
+  - @globalfishingwatch/timebar@1.6.1
+
+## 0.7.84
+
+### Patch Changes
+
+- 1a3d0960: Fishing map/presence less than one day
+- Updated dependencies [1a3d0960]
+  - @globalfishingwatch/layer-composer@7.0.0
+  - @globalfishingwatch/timebar@1.6.0
+  - @globalfishingwatch/react-hooks@7.0.0
+
+## 0.7.83
+
+### Patch Changes
+
+- Updated dependencies [7d060771]
+  - @globalfishingwatch/layer-composer@6.1.4
+  - @globalfishingwatch/react-hooks@6.3.1
+
 ## 0.7.82
 
 ### Patch Changes